+++ packages/web/pages/api/router/token-out-given-in.ts
@@ -2,7 +2,7 @@ import { Int } from "@keplr-wallet/unit";
 import { Route, SplitTokenInQuote } from "@osmosis-labs/pools";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { routeTokenOutGivenIn as _routeTokenOutGivenIn } from "~/server/queries/complex/route-token-out-given-in";
+import { routeTokenOutGivenIn as _routeTokenOutGivenIn } from "~/server/queries/complex/pools/route-token-out-given-in";
 
 type Response = {
   amount: string;
