+++ packages/web/server/queries/complex/assets/index.ts
@@ -1,4 +1,3 @@
-import { CoinPretty, Dec, PricePretty, RatePretty } from "@keplr-wallet/unit";
 import { AssetList } from "@osmosis-labs/types";
 import { makeMinimalAsset } from "@osmosis-labs/utils";
 import cachified, { CacheEntry } from "cachified";
@@ -8,13 +7,6 @@ import { z } from "zod";
 import { DEFAULT_LRU_OPTIONS } from "~/config/cache";
 import { AssetLists } from "~/config/generated/asset-lists";
 import { search, SearchSchema } from "~/utils/search";
-import { SortDirection } from "~/utils/sort";
-
-import { queryBalances } from "../../cosmos";
-import { queryTokenMarketCaps } from "../../indexer";
-import { DEFAULT_VS_CURRENCY } from "./config";
-import { getAssetData, getAssetMarketCapRank } from "./info";
-import { calcAssetValue, getAssetPrice } from "./price";
 
 /** An asset with minimal data that conforms to `Currency` type. */
 export type Asset = {
@@ -66,231 +58,73 @@ export async function getAssets({
   /** Explicitly match the base or symbol denom. */
   findMinDenomOrSymbol?: string;
 } & AssetFilter = {}): Promise<Asset[]> {
-  const makeMinimalAssets = (assetList: AssetList[]) => {
-    // Create new array with just assets
-    const coinMinimalDenomSet = new Set<string>();
-
-    const listedAssets = assetList
-      .flatMap(({ assets }) => assets)
-      .filter(
-        (asset) =>
-          asset.keywords && !asset.keywords.includes("osmosis-unlisted")
-      );
-
-    let assetListAssets = listedAssets.filter((asset) => {
-      if (params.findMinDenomOrSymbol) {
-        return (
-          params.findMinDenomOrSymbol.toUpperCase() ===
-            asset.base.toUpperCase() ||
-          params.findMinDenomOrSymbol.toUpperCase() ===
-            asset.symbol.toUpperCase()
-        );
-      }
-
-      // Ensure denoms are unique on Osmosis chain
-      // In the case the asset list has the same asset twice
-      if (coinMinimalDenomSet.has(asset.base)) {
-        return false;
-      } else {
-        coinMinimalDenomSet.add(asset.base);
-        return true;
-      }
-    });
-
-    // Search raw asset list before reducing type to minimal Asset type
-    if (params.search) {
-      assetListAssets = search(
-        assetListAssets,
-        searchableAssetListAssetKeys,
-        params.search
-      );
-    }
-
-    // Filter by only verified
-    if (params.onlyVerified) {
-      assetListAssets = assetListAssets.filter((asset) =>
-        Boolean(asset.keywords?.includes("osmosis-main"))
-      );
-    }
-
-    // Transform into a more compact object
-    return assetListAssets.map(makeMinimalAsset);
-  };
-
   // if it's the default asset list, cache it
   if (assetList === AssetLists) {
     return cachified({
       cache: minimalAssetsCache,
       key: JSON.stringify(params),
-      getFreshValue: () => makeMinimalAssets(assetList),
+      getFreshValue: () => reduceAssetList(assetList, params),
     });
   }
 
   // otherwise process the given novel asset list
-  return makeMinimalAssets(assetList);
+  return reduceAssetList(assetList, params);
 }
 
-/** Basic user info if the user holds a balance. */
-export type MaybeUserAssetInfo = Partial<{
-  amount: CoinPretty;
-  usdValue: PricePretty;
-}>;
+/** Transform given asset list into an array of minimal asset types for user in frontend. */
+function reduceAssetList(
+  assetList: AssetList[],
+  params: { findMinDenomOrSymbol?: string } & AssetFilter = {}
+): Asset[] {
+  // Create new array with just assets
+  const coinMinimalDenomSet = new Set<string>();
 
-export async function getUserAssetInfo<TAsset extends Asset>({
-  assetList = AssetLists,
-  asset,
-  userOsmoAddress,
-}: {
-  assetList?: AssetList[];
-  asset: TAsset;
-  userOsmoAddress?: string;
-}): Promise<TAsset & MaybeUserAssetInfo> {
-  if (!userOsmoAddress) return asset;
+  const listedAssets = assetList
+    .flatMap(({ assets }) => assets)
+    .filter(
+      (asset) => asset.keywords && !asset.keywords.includes("osmosis-unlisted")
+    );
 
-  const userAssets = await mapGetUserAssetInfos({
-    assetList,
-    assets: [asset],
-    userOsmoAddress,
+  let assetListAssets = listedAssets.filter((asset) => {
+    if (params.findMinDenomOrSymbol) {
+      return (
+        params.findMinDenomOrSymbol.toUpperCase() ===
+          asset.base.toUpperCase() ||
+        params.findMinDenomOrSymbol.toUpperCase() === asset.symbol.toUpperCase()
+      );
+    }
+
+    // Ensure denoms are unique on Osmosis chain
+    // In the case the asset list has the same asset twice
+    if (coinMinimalDenomSet.has(asset.base)) {
+      return false;
+    } else {
+      coinMinimalDenomSet.add(asset.base);
+      return true;
+    }
   });
-  return userAssets[0];
-}
 
-/** Maps user asset balance data given a list of assets of a given type and a potential user Osmosis address.
- *  If no assets provided, they will be fetched and passed the given search params.
- *  If no search param is provided and `sortFiatValueDirection` is defined, it will sort by user fiat value.  */
-export async function mapGetUserAssetInfos<TAsset extends Asset>({
-  assetList = AssetLists,
-  ...params
-}: {
-  assetList?: AssetList[];
-  assets?: TAsset[];
-  userOsmoAddress?: string;
-  sortFiatValueDirection?: SortDirection;
-} & AssetFilter): Promise<(TAsset & MaybeUserAssetInfo)[]> {
-  const { userOsmoAddress, search, sortFiatValueDirection } = params;
-  let { assets } = params;
-  if (!assets) assets = (await getAssets(params)) as TAsset[];
-  if (!userOsmoAddress) return assets;
-
-  const { balances } = await queryBalances(userOsmoAddress);
-
-  const eventualUserAssets = assets
-    .map(async (asset) => {
-      const balance = balances.find((a) => a.denom === asset.coinMinimalDenom);
-
-      // not a user asset
-      if (!balance) return asset;
-
-      // is user asset, include user data
-      const usdValue = await calcAssetValue({
-        anyDenom: asset.coinMinimalDenom,
-        amount: balance.amount,
-      }).catch(() => {
-        console.error(asset.coinMinimalDenom, "likely missing price config");
-      });
-
-      return {
-        ...asset,
-        amount: new CoinPretty(asset, balance.amount),
-        usdValue: usdValue
-          ? new PricePretty(DEFAULT_VS_CURRENCY, usdValue)
-          : undefined,
-      };
-    })
-    .filter((a): a is Promise<TAsset & MaybeUserAssetInfo> => !!a);
-
-  const userAssets = await Promise.all(eventualUserAssets);
-
-  // if no search provided, sort by usdValue at head of list by default
-  if (!search && sortFiatValueDirection) {
-    userAssets.sort((a, b) => {
-      if (!Boolean(a.usdValue) && !Boolean(b.usdValue)) return 0;
-      if (Boolean(a.usdValue) && !Boolean(b.usdValue)) return -1;
-      if (!Boolean(a.usdValue) && Boolean(b.usdValue)) return 1;
-
-      if (sortFiatValueDirection === "desc") {
-        const n = Number(
-          b.usdValue!.toDec().sub(a.usdValue!.toDec()).toString()
-        );
-        if (isNaN(n)) return 0;
-        else return n;
-      } else {
-        const n = Number(
-          a.usdValue!.toDec().sub(b.usdValue!.toDec()).toString()
-        );
-        if (isNaN(n)) return 0;
-        else return n;
-      }
-    });
+  // Search raw asset list before reducing type to minimal Asset type
+  if (params.search) {
+    assetListAssets = search(
+      assetListAssets,
+      searchableAssetListAssetKeys,
+      params.search
+    );
   }
 
-  return userAssets;
+  // Filter by only verified
+  if (params.onlyVerified) {
+    assetListAssets = assetListAssets.filter((asset) =>
+      Boolean(asset.keywords?.includes("osmosis-main"))
+    );
+  }
+
+  // Transform into a more compact object
+  return assetListAssets.map(makeMinimalAsset);
 }
 
-export type AssetMarketInfo = Partial<{
-  marketCap: PricePretty;
-  marketCapRank: number;
-  currentPrice: PricePretty;
-  priceChange24h: RatePretty;
-}>;
-
-/** Maps and adds general supplementary market data such as current price and market cap to the given type.
- *  If no assets provided, they will be fetched and passed the given search params. */
-export async function mapGetAssetMarketInfos<TAsset extends Asset>({
-  assetList = AssetLists,
-  ...params
-}: {
-  assetList?: AssetList[];
-  assets?: TAsset[];
-} & AssetFilter): Promise<(TAsset & AssetMarketInfo)[]> {
-  let { assets } = params;
-  if (!assets) assets = (await getAssets(params)) as TAsset[];
-
-  return await Promise.all(
-    assets.map((asset) => getAssetMarketInfo({ asset }))
-  );
-}
-
-const marketInfoCache = new LRUCache<string, CacheEntry>(DEFAULT_LRU_OPTIONS);
-/** Cached function that returns an asset with market info included. */
-export async function getAssetMarketInfo<TAsset extends Asset>({
-  asset,
-}: {
-  asset: TAsset;
-}): Promise<TAsset & AssetMarketInfo> {
-  return cachified({
-    cache: marketInfoCache,
-    key: asset.coinDenom + asset.coinMinimalDenom,
-    ttl: 1000 * 60 * 5, // 5 minutes
-    getFreshValue: async () => {
-      const currentPrice = await getAssetPrice({ asset }).catch(() => {
-        // if not found, return undefined
-        return;
-      });
-      const marketCap = (
-        await queryTokenMarketCaps().catch(() => {
-          // if not found, return undefined
-          return undefined;
-        })
-      )?.find((mCap) => mCap.symbol === asset.coinDenom)?.market_cap;
-      const priceChange24h = (await getAssetData(asset))?.price_24h_change;
-      const marketCapRank = await getAssetMarketCapRank(asset);
-
-      return {
-        ...asset,
-        currentPrice: currentPrice
-          ? new PricePretty(DEFAULT_VS_CURRENCY, currentPrice)
-          : undefined,
-        marketCap: marketCap
-          ? new PricePretty(DEFAULT_VS_CURRENCY, marketCap)
-          : undefined,
-        marketCapRank,
-        priceChange24h: priceChange24h
-          ? new RatePretty(new Dec(priceChange24h).quo(new Dec(100)))
-          : undefined,
-      };
-    },
-  });
-}
-
+export * from "./info";
+export * from "./market";
 export * from "./price";
+export * from "./user";
