+++ packages/web/server/queries/complex/pools/route-token-out-given-in.ts
@@ -22,8 +22,8 @@ import { LRUCache } from "lru-cache";
 import { DEFAULT_LRU_OPTIONS } from "~/config/cache";
 import { ChainList } from "~/config/generated/chain-list";
 
-import { queryNumPools } from "../osmosis";
-import { queryPaginatedPools } from "./pools/providers/indexer";
+import { queryNumPools } from "../../osmosis";
+import { queryPaginatedPools } from "./providers/indexer";
 
 /**
  * This function routes a given token to a specified output token denomination.
